# Conflicts:
#	pyleecan/Classes/Class_Dict.json
#	pyleecan/Classes/ConvertMC.py
#	pyleecan/Generator/ClassesRef/Converter/ConvertMC.csv
#	pyleecan/Methods/Slot/SlotW63/build_geometry_active.py

+++ pyleecan/Methods/Slot/SlotW62/comp_angle_opening.py
@@ -18,6 +18,6 @@ def comp_angle_opening(self):
 
     """
     slot_pitch = 2 * pi / self.Zs
-    tooth_angle = float(2 * arcsin(self.W1 / (2 * self.get_Rbo())))
+    tooth_angle = 2 * arcsin(self.W1 / (2 * self.get_Rbo()))
 
     return slot_pitch - tooth_angle
